refactor(server): split main.py and fragments.py into focused modules
main.py had grown to 1740 lines (one closures-based create_app() factory
holding ~50 routes) and fragments.py to 1019 lines of flat HTML-template
functions. Split both along the app's actual resource boundaries — health,
sessions, streaming, admin status/tokens/models/config, pairing — using
FastAPI's Depends(get_context) DI pattern instead of closures over shared
state. Every file now lands under ~300 lines.
