+++ server/app/fragments/__init__.py
@@ -1,0 +1,1 @@
+"""HTML fragment templates for the admin WebUI (HTMX partials)."""
