+++ server/app/routes/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI routers, one module per resource area of the gateway."""
